note: Move the instrument register to note_initialize
avoid the instrument doesn't get registered in some case
