+++ drivers/note/note_driver.c
@@ -41,7 +41,6 @@
 #include <nuttx/note/notestream_driver.h>
 #include <nuttx/spinlock.h>
 #include <nuttx/sched_note.h>
-#include <nuttx/instrument.h>
 
 #include "sched/sched.h"
 #include "noterpmsg.h"
@@ -156,18 +155,6 @@ struct note_taskname_s
 /****************************************************************************
  * Private Data
  ****************************************************************************/
-
-#ifdef CONFIG_SCHED_INSTRUMENTATION_FUNCTION
-static void note_driver_instrument_enter(FAR void *this_fn,
-            FAR void *call_site, FAR void *arg) noinstrument_function;
-static void note_driver_instrument_leave(FAR void *this_fn,
-            FAR void *call_site, FAR void *arg) noinstrument_function;
-static struct instrument_s g_note_instrument =
-{
-  .enter = note_driver_instrument_enter,
-  .leave = note_driver_instrument_leave,
-};
-#endif
 
 #ifdef CONFIG_SCHED_INSTRUMENTATION_FILTER
 static struct note_filter_s g_note_filter =
@@ -1997,24 +1984,6 @@ FAR const char *note_get_taskname(pid_t pid)
 
 #endif
 
-#ifdef CONFIG_SCHED_INSTRUMENTATION_FUNCTION
-static void note_driver_instrument_enter(FAR void *this_fn,
-                                         FAR void *call_site,
-                                         FAR void *arg)
-{
-  sched_note_event_ip(NOTE_TAG_ALWAYS, (uintptr_t)this_fn,
-                      NOTE_DUMP_BEGIN, NULL, 0);
-}
-
-static void note_driver_instrument_leave(FAR void *this_fn,
-                                         FAR void *call_site,
-                                         FAR void *arg)
-{
-  sched_note_event_ip(NOTE_TAG_ALWAYS, (uintptr_t)this_fn,
-                      NOTE_DUMP_END, NULL, 0);
-}
-#endif
-
 /****************************************************************************
  * Name: note_driver_register
  ****************************************************************************/
@@ -2022,15 +1991,6 @@ static void note_driver_instrument_leave(FAR void *this_fn,
 int note_driver_register(FAR struct note_driver_s *driver)
 {
   int i;
-#ifdef CONFIG_SCHED_INSTRUMENTATION_FUNCTION
-  static bool initialized;
-
-  if (!initialized)
-    {
-      instrument_register(&g_note_instrument);
-      initialized = true;
-    }
-#endif
 
   DEBUGASSERT(driver);
   for (i = 0; i < CONFIG_DRIVERS_NOTE_MAX; i++)
